requirements.txt
appdirs==1.4.3
astroid==2.4.2
attrs==20.2.0
backcall==0.2.0
CacheControl==0.12.6
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2
django-appconf==1.0.4
django-braces==1.14.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-etc==1.2.0
django-hitcount==1.3.3
django-jalali-date==0.3.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-redis==4.12.1
django-widget-tweaks==1.4.8
html5lib==1.0.1
humanize==3.2.0
idna==2.8
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jdatetime==3.6.2
jedi==0.18.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
ptyprocess==0.7.0
Pygments==2.7.4
pylint==2.5.3
pyparsing==2.4.6
python-decouple==3.3
python-magic==0.4.18
pytoml==0.1.21
pytz==2020.1
redis==3.5.3
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zeep==3.4.0
